updated frontend and backend: frontend 65%; backend 95%

+++ backend/modules/blogPost/blogPost.service.js
@@ -1,11 +1,13 @@
 const UserSchema = require('../user/user.schema')
 const BlogPostSchema = require('./blogPost.schema')
+const CommentSchema = require('../comment/comment.schema')
 
 const getBlogPosts = async (page, pageSize) => {
   const blogPosts = await BlogPostSchema.find()
     .limit(pageSize)
     .skip((page - 1) * pageSize)
     .populate('user', 'name surname avatar nationality isTutor')
+    .populate('comments')
   const totalBlogPosts = await BlogPostSchema.countDocuments()
   const totalPages = Math.ceil(totalBlogPosts / pageSize)
   return {
@@ -22,7 +24,7 @@ const getBlogPostsByTitle = async (searchedTitle, page, pageSize) => {
   const blogPosts = await BlogPostSchema.find(searchQuery)
     .limit(pageSize)
     .skip((page - 1) * pageSize)
-    .populate('user', 'name surname avatar')
+    .populate('user', 'name surname avatar nationality isTutor')
   const totalBlogPosts = await BlogPostSchema.find(searchQuery).countDocuments()
   const totalPages = Math.ceil(totalBlogPosts / pageSize)
   return {
@@ -36,6 +38,13 @@ const getBlogPostsByTitle = async (searchedTitle, page, pageSize) => {
 
 const getBlogPostById = async (blogPostId) => {
   const blogPost = await BlogPostSchema.findById(blogPostId).populate('user', 'name surname avatar nationality isTutor')
+    .populate({
+      path: 'comments',
+      populate: {
+        path: 'user',
+        select: 'name surname avatar nationality isTutor'
+      }
+    });
   return blogPost
 }
 
@@ -53,8 +62,41 @@ const updateBlogPost = async (blogPostId, body) => {
 }
 
 const deleteBlogPost = async (blogPostId) => {
-  return await BlogPostSchema.findByIdAndDelete(blogPostId)
+  const post = await BlogPostSchema.findById(blogPostId);
+  if (post) {
+    await UserSchema.updateOne({ _id: post.user }, { $pull: { blogPosts: blogPostId } })
+    await CommentSchema.deleteMany({ blogPost: blogPostId })
+    return await BlogPostSchema.findByIdAndDelete(blogPostId)
+  }
+  return null;
 }
+
+const toggleLike = async (blogPostId, userId) => {
+  const blogPost = await BlogPostSchema.findById(blogPostId);
+  if (!blogPost) throw new Error("Post non trovato");
+
+  const hasLiked = blogPost.likes.includes(userId);
+  const update = hasLiked 
+    ? { $pull: { likes: userId } } 
+    : { $addToSet: { likes: userId } };
+
+  return await BlogPostSchema.findByIdAndUpdate(
+    blogPostId,
+    update,
+    { new: true }
+  )
+  .populate('user', 'name surname avatar nationality isTutor')
+  .populate({
+    path: 'comments',
+    populate: {
+      path: 'user',
+      select: 'name surname avatar nationality isTutor'
+    }
+  });
+};
+
+
+
 
 module.exports = {
   getBlogPosts,
@@ -62,5 +104,6 @@ module.exports = {
   getBlogPostById,
   createBlogPost,
   updateBlogPost,
-  deleteBlogPost
+  deleteBlogPost,
+  toggleLike
 }
